pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-layout-translator"
version = "0.2.0"
description = "Telegram-first low-cost PDF translation API with OpenRouter"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "pymupdf>=1.24.0",
  "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.5.6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
